Separate deserialization problem types in two
Scenario:
Consumer makes a request to the Release/Details endpoint.
Response is not really a Release/Details response.
It is still XML, and it has a <response> element, and an "ok" status,
but it is something else that cannot be deserialized to a
T==typeof(ReleaseDetails).

Previous results:
We were getting an extraneous "NonXmlResponseException", but the response
was clearly XML, so we have been some days trying to dig if the servers
where sending extraneous characters or BOMs that could not be well represented
when printing the ResponseBody in the stacktrace.

The real culprit of this bug we're getting is obviously not fixed by this
pull-request (we think it's a threading issue, more on that later), but at
least with this changeSet one could be not so confused!
Thanks

Conflicts:

	src/SevenDigital.Api.Wrapper/SevenDigital.Api.Wrapper.csproj